% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{CI}
\alias{CI}
\title{Confidence interval}
\usage{
CI(x, level = 0.95, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an \R object which is coercible to one by
\code{as.vector(x, "numeric")}.}

\item{level}{the confidence level required.}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
This wrapper function computes the confidence interval of the values in \code{x}.
If \code{na.rm} is \code{TRUE}, missing values are removed before the 
computation proceeds.
}
\examples{
CI(1:5)

}
\seealso{
\code{\link[stats]{sd}} for the standard deviation.
}
