% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clayloam.R
\docType{data}
\name{clayloam}
\alias{clayloam}
\title{Hydrometer readings for a clay loam}
\format{
A data frame containing 7 rows and 4 columns with information on:

\describe{
  \item{time}{the measurement time [min]}
  \item{temperature}{the temperature of the soil suspension [°C]}
  \item{reading}{the hydrometer reading at the bottom of the meniscus}
  \item{blank}{a blank value obtained in 5 g/L sodium hexametaphosphate
  solution (composite correction)}
}
}
\description{
Data obtained with a 152H hydrometer in accordance with ASTM D422-63 (2007).
}
\references{
ASTM D422-63 (2007). \emph{Standard Test Method for Particle-Size Analysis
of Soils}. Technical standard. ASTM International, West Conshohocken, PA.
Available from \url{https://www.astm.org/standards/d422}.
}
\seealso{
Other texture: 
\code{\link{texture}()}
}
\author{
ASTM International
}
\concept{texture}
\keyword{data}
