% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/din32645.R
\docType{data}
\name{din32645}
\alias{din32645}
\title{Calibration data from DIN 32645}
\format{
A data frame containing 20 rows and 2 columns with information on:
\tabular{ll}{
Conc: \tab nominal concentration in mg/L\cr
Area: \tab measurement signal\cr
}
}
\description{
Sample data for the calibration of carbon in water.
}
\references{
DIN 32645 (2008). \emph{Chemical analysis - Decision limit, detection limit
and determination limit under repeatability conditions - Terms, methods,
evaluation}. Technical standard. Deutsches Institut für Normung, Berlin.
}
\seealso{
Other calibration: 
\code{\link{calibration}()},
\code{\link{icp}},
\code{\link{matrix_effect}()},
\code{\link{neitzel2003}},
\code{\link{weight_select}()}
}
\author{
Deutsches Institut für Normung
}
\concept{calibration}
\keyword{data}
