% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envalysis-package.R
\docType{package}
\name{envalysis-package}
\alias{envalysis}
\alias{envalysis-package}
\title{envalysis: Miscellaneous Functions for Environmental Analyses}
\description{
Small toolbox for data analyses in environmental chemistry and ecotoxicology. Provides, for example, calibration() to calculate calibration curves and corresponding limits of detection (LODs) and limits of quantification (LOQs) according to German DIN 32645 (2008). texture() makes it easy to estimate soil particle size distributions from hydrometer measurements (ASTM D422-63, 2007).
}
\references{
Almeida, A. M. D., Castel-Branco, M. M., & Falcao, A. C. (2002). Linear
regression for calibration lines revisited: weighting schemes for
bioanalytical methods. \emph{Journal of Chromatography B}, \bold{774}(2),
215-222. \doi{10.1016/S1570-0232(02)00244-1}.

ASTM D422-63 (2007). \emph{Standard Test Method for Particle-Size Analysis
of Soils}. Technical standard. ASTM International, West Conshohocken, PA.
Available from \url{https://www.astm.org/standards/d422}.

Borcard D, Gillet F, & Legendre P. (2011). \emph{Numerical ecology with R}.
Springer, New York.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zsteinmetz/envalysis}
  \item \url{https://zsteinmetz.de/envalysis/}
  \item Report bugs at \url{https://github.com/zsteinmetz/envalysis/issues}
}

}
\author{
\strong{Maintainer}: Zacharias Steinmetz \email{z.steinmetz@rptu.de} (\href{https://orcid.org/0000-0001-6675-5033}{ORCID})

Other contributors:
\itemize{
  \item Julius Albert (\href{https://orcid.org/0000-0001-5468-6069}{ORCID}) [contributor]
  \item Kilian Kenngott (\href{https://orcid.org/0000-0002-9600-4502}{ORCID}) [contributor]
}

}
\keyword{internal}
