% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{envalysis}
\alias{envalysis}
\alias{envalysis-package}
\title{Miscellaneous functions for environmental analyses}
\description{
Data manipulation tools and functions that may be helpful for data analyses in
environmental chemistry and ecotoxicology.
}
\details{
\tabular{ll}{
Version: \tab 0.1\cr
Licence: \tab GPLv3\cr
}
}
\author{
Zacharias Steinmetz (\email{stei4785@uni-landau.de})
}
\references{
Borcard D, Gillet F, Legendre P. 2011. Numerical ecology with R. Springer, New York.
}

