% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/icp.R
\docType{data}
\name{icp}
\alias{icp}
\title{ICP-AES calibration data}
\format{A data frame containing 16 rows and 6 columns with information on:
\tabular{ll}{
Label: \tab sample name\cr
Element: \tab element name\cr
Conc: \tab nominal concentration\cr
Units: \tab concentration units\cr
Signal: \tab signal intensity [cps]\cr
Replicate: \tab number of the respective replicate\cr
}}
\usage{
data(icp)
}
\description{
Data on the calibration of silver ions in four-fold replication using
inductively coupled plasma atomic emission spectroscopy (ICP-AES).
}
\keyword{data}

