% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loq.R
\name{loq}
\alias{loq}
\title{Limit of quantification (LOQ)}
\usage{
loq(object, conc, alpha = 0.05, beta = 0.5)
}
\arguments{
\item{object}{a univariate model object of class \code{lm} with a model formula
\code{Signal ~ Conc} or \code{Signal ~ Conc - 1} containing the linear regression of the
analyte concentrations with respect to the signal intensity recorded}

\item{conc}{the column expression specifying the respective concentrations}

\item{alpha}{the error tolerance for the quantification limit (critical value)}

\item{beta}{the error tolerance for a false negative}
}
\description{
Estimates the LOQ based on a given calibration function. The LOQ is defined as the lowest
quantity of a substance that can be quantified / distinguished from another sample
with confidence.
}
\examples{
\dontrun{
data(icp)
Cal <- lm(Signal ~ Conc, data = icp)
loq(Cal, Conc)
}
}
\references{
Currie, LA. 1997. Nomenclature in evaluation of analytical methods including detection
and quantification capabilities (IUPAC Recommendations 1995). Analytica Chimica Acta 391:105-126.
}

