% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.raw.R
\name{make.raw}
\alias{make.raw}
\title{Convert frequencies to raw data}
\usage{
make.raw(data, var.name, freq.name)
}
\arguments{
\item{data}{a data frame or matrix containing the frequency information}

\item{var.name}{a character string specifying the column with the respective variable}

\item{freq.name}{a character string specifying the column with the respective frequency information}
}
\description{
Convert a data frame or matrix containing summarized frequencies information back
to raw data.
}
\examples{
data <- data.frame(var = c(10.2,15,5), freq=c(7,6,10))
make.raw(data, "var", "freq")

}
\seealso{
\code{\link[graphics]{hist}}
}

