% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_effect.R
\name{matrix_effect}
\alias{matrix_effect}
\title{Assess matrix effects and matrix-matched calibrations}
\usage{
matrix_effect(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link[envalysis]{calibration}}
obtained by analyzing standard solutions.}

\item{\dots}{additional objects of the same type obtained from
matrix matched calibrations.}
}
\description{
Assess matrix effects and matrix-matched calibrations
}
\examples{
data(din32645)
din <- calibration(Area ~ Conc, data = din32645)

m32645 <- din32645
m32645$Area <- din32645$Area * 1.5
matrix <- calibration(Area ~ Conc, data = m32645)

matrix_effect(din, matrix)

}
\seealso{
Other calibration: 
\code{\link{calibration}()},
\code{\link{weight_select}()}
}
\concept{calibration}
