% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neitzel2003.R
\docType{data}
\name{neitzel2003}
\alias{neitzel2003}
\title{Calibration data from Neitzel, 2003}
\format{
A data frame containing 20 rows and 2 columns with information on:
\tabular{ll}{
Conc: \tab nominal concentration\cr
Meas: \tab measurement signal\cr
}
}
\description{
Artificial sample data for the verification of quantification limits.
}
\references{
Neitzel, V., 2003. Kalibrierung bei Analysenverfahren - Bestimmungsgrenze ist
nicht gleich Bestimmungsgrenze. CLB Chemie in Labor und Biotechnik, 54(7),
242–246
}
\seealso{
\code{\link{calibration}}
}
\keyword{data}
