% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme_publish.R
\name{theme_publish}
\alias{theme_publish}
\title{ggplot2 theme for scientific publications}
\usage{
theme_publish(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Themes set the general aspect of the plot such as the color of the
background, gridlines, the size and colour of fonts. This particular theme is
based on the classic dark-on-light ggplot2 \code{theme_bw} and has been used
for scientific publications,
}
\examples{
require(ggplot2)
require(grid)
p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,
     colour=factor(gear))) + facet_wrap( ~ am)
p
p + theme_publish()
}
\seealso{
\code{\link[ggplot2]{ggtheme}}
}

