% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_select.R
\name{weight_select}
\alias{weight_select}
\alias{weight_select.calibration}
\alias{relerr}
\alias{relerr.calibration}
\title{Tools for weighted calibrations}
\usage{
weight_select(object, weights)

\method{weight_select}{calibration}(object, weights)

relerr(object)

\method{relerr}{calibration}(object)
}
\arguments{
\item{object}{an object of class \code{\link[envalysis]{calibration}}.}

\item{weights}{a list of weights.}
}
\description{
Tools for weighted calibrations
}
\examples{
data(din32645)
din <- calibration(Area ~ Conc, data = din32645)

relerr(din)

}
\seealso{
Other calibration: 
\code{\link{calibration}()},
\code{\link{matrix_effect}()}
}
\concept{calibration}
