% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\alias{lod}
\alias{loq}
\alias{summary.calibration}
\title{Analytical calibration functions}
\usage{
calibration(formula, data, model = "lm", ...)

\method{summary}{calibration}(object, ...)

lod(object, alpha = 0.01)

loq(object, alpha = 0.01, k = 3)
}
\arguments{
\item{formula}{model formula providing the recorded signal intensities with
respect to the nominal analyte concentrations in the form of
\code{signal ~ concentration} or \code{signal ~ concentraion - 1}}

\item{data}{an optional data frame containing the variables in the model}

\item{model}{model class to be used for fitting; currently, only \code{lm}
is supported}

\item{object}{a univariate model object of class \code{calibration} or \code{lm}
with a model formula as shown above}

\item{alpha}{error tolerance for the detection limit (critical value)}

\item{k}{relative uncertainty for the limit of quantification (1/beta)}

\item{\dots}{further arguments passed to the model environment}
}
\value{
\code{calibration} returns an object of \code{\link[base]{class}}
"calibration". \code{summary} calls the model summary together with the
respective LOD and LOQ.
}
\description{
Defines a \code{calibration} object for the calculation of concentrations
from measurement signals including estimatations for the limit of detection
(LOD) and limit of quantification (LOQ) in accordance with DIN
32645:2008-11.

The LOD is defined as the lowest quantity of a substance that can be
distinguished from the absence of that substance (blank value) within a
given confidence level (alpha). The LOQ is defined as the lowest quantity of
a substance that can be quantified/distinguished from another sample given
with respect to a defined confidence level (k).
}
\examples{
data(din32645)
din <- calibration(Area ~ Conc, data = din32645)
summary(din)

lod(din)

loq(din)

}
\references{
DIN 32645:2008-11, 2008. Chemical analysis - Decision limit, detection limit
and determination limit under repeatability conditions - Terms, methods,
evaluation (Technical standard). Deutsches Institut für Normung, Berlin.

Currie, L.A., 1999. Nomenclature in evaluation of analytical methods
including detection and quantification capabilities:
(IUPAC Recommendations 1995). Analytica Chimica Acta 391, 105–126.

Massart, D.L., Vandeginste, B.G., Buydens, L.M.C., Lewi, P.J.,
Smeyers-Verbeke, J., 1997. Handbook of chemometrics and qualimetrics:
Part A. Elsevier Science Inc.
}
\seealso{
\code{\link{icp}}, \code{\link{din32645}}
}

