% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/din32645.R
\docType{data}
\name{din32645}
\alias{din32645}
\title{Calibration data from DIN 32645}
\format{A data frame containing 20 rows and 2 columns with information on:
\tabular{ll}{
Conc: \tab nominal concentration in mg/L\cr
Area: \tab measurement signal\cr
}}
\usage{
data(din32645)
}
\description{
Sample data for the calibration of carbon in water.
}
\seealso{
\code{\link{calibration}}
}
\keyword{data}

