% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{envalysis}
\alias{envalysis}
\alias{envalysis-package}
\title{Miscellaneous Functions for Environmental Analyses}
\description{
Data manipulation tools and functions that may be helpful for data analyses in
environmental chemistry and ecotoxicology.
}
\details{
\tabular{ll}{
Version: \tab 0.3.2\cr
Licence: \tab GPLv3\cr
}
}
\references{
Borcard D, Gillet F, Legendre P. 2011. Numerical ecology with R. Springer, New York.
}
\author{
Zacharias Steinmetz (\email{steinmetz-z@uni-landau.de})
}
