% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{envalysis}
\alias{envalysis}
\title{Miscellaneous Functions for Environmental Analyses}
\description{
Small toolbox for data analyses in environmental chemistry and
ecotoxicology. Provides, for example, calibration() to calculate calibration
curves and corresponding limits of detection (LODs) and quantification
(LOQs) according to German DIN 32645:2008-11. texture() makes it easy to
estimate soil particle size distributions from hydrometer measurements (ASTM
D422-63(2007)e2).
}
\details{
\tabular{ll}{
Version: \tab 0.4.2\cr
Licence: \tab GPLv3\cr
}
}
\references{
Borcard D, Gillet F, Legendre P. 2011. Numerical ecology with R. Springer,
New York.
}
\author{
Zacharias Steinmetz (\email{steinmetz-z@uni-landau.de})
}
