% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icp.R
\docType{data}
\name{icp}
\alias{icp}
\title{ICP-AES calibration data}
\format{
A data frame containing 16 rows and 6 columns with information on:

\describe{
  \item{Label}{sample name}
  \item{Element}{element name}
  \item{Conc}{nominal concentration}
  \item{Units}{concentration units}
  \item{Signal}{signal intensity [cps]}
  \item{Replicate}{number of the respective replicate}
}
}
\description{
Sample data of the calibration of silver ions in four-fold replication using
inductively coupled plasma atomic emission spectroscopy (ICP-AES).
}
\seealso{
Other calibration: 
\code{\link{calibration}()},
\code{\link{din32645}},
\code{\link{matrix_effect}()},
\code{\link{neitzel2003}},
\code{\link{phenolics}},
\code{\link{weight_select}()}
}
\author{
Zacharias Steinmetz
}
\concept{calibration}
\keyword{data}
