% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lod.R
\name{lod}
\alias{lod}
\title{Limit of detection (LOD)}
\usage{
lod(object, conc, alpha = 0.05)
}
\arguments{
\item{object}{a univariate model object of class \code{lm} with a model formula
\code{Signal ~ Conc} or \code{Signal ~ Conc - 1} containing the linear regression of the
analyte concentrations with respect to the signal intensity recorded}

\item{conc}{the column expression specifying the respective concentrations}

\item{alpha}{the error tolerance for the detection limit (critical value)}
}
\description{
Estimates the LOD based on the calibration function. The LOD is defined as the lowest
quantity of a substance that can be distinguished from the absence of that substance
(blank value) within a stated confidence level.
}
\examples{
\dontrun{
data(icp)
Cal <- lm(Signal ~ Conc, data = icp)
lod(Cal, Conc)
}
}
\references{
Currie, LA. 1997. Nomenclature in evaluation of analytical methods including detection
and quantification capabilities (IUPAC Recommendations 1995). Analytica Chimica Acta 391:105-126.
}

