% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neitzel2003.R
\docType{data}
\name{neitzel2003}
\alias{neitzel2003}
\title{Calibration data from Neitzel, 2003}
\format{
A data frame containing 20 rows and 2 columns with information on:
\tabular{ll}{
Conc: \tab nominal concentration\cr
Meas: \tab measurement signal\cr
}
}
\description{
Artificial sample data for the verification of quantification limits.
}
\references{
Neitzel, V. (2003). Kalibrierung bei Analysenverfahren - Bestimmungsgrenze ist
nicht gleich Bestimmungsgrenze. \emph{CLB Chemie in Labor und Biotechnik},
\bold{54}(7), 242-246
}
\seealso{
Other calibration: 
\code{\link{calibration}()},
\code{\link{din32645}},
\code{\link{icp}},
\code{\link{matrix_effect}()},
\code{\link{phenolics}},
\code{\link{weight_select}()}
}
\author{
Volkmar Neitzel
}
\concept{calibration}
\keyword{data}
