% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/puri.test.R
\name{puri.test}
\alias{puri.test}
\title{Sen–Puri test}
\usage{
puri.test(formula, data = NULL, ...)
}
\arguments{
\item{formula}{a formula specifying the model.}

\item{data}{data frame in which the variables specified in the formula will be found}
}
\description{
Performs an analysis of variance (ANOVA) on ranks according to Sen and Puri.
}
\examples{
puri.test(yield ~ N + P + K, data = npk)
}
\references{
Puri ML, Sen PK. 1969. A Class of Rank Order Tests for a General Linear Hypothesis.
The Annals of Mathematical Statistics. 40:1325–1343.
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[stats]{aov}}
}

