% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puri.test.R
\name{puri.test}
\alias{puri.test}
\title{Sen-Puri test}
\usage{
puri.test(formula, data, ...)
}
\arguments{
\item{formula}{a formula specifying the model.}

\item{data}{data frame in which the variables specified in the formula will be found}

\item{\dots}{further arguments to be passed to \code{lm}, such as \code{subset} or
\code{na.action}}
}
\description{
Performs an analysis of variance (ANOVA) on ranks according to
Sen and Puri (1969), also known as Scheirer-Ray-Hare-Test (Scheirer et al.,
1976)
}
\examples{
data(npk)
puri.test(yield ~ N + P + K, data = npk)

}
\references{
Puri, M.L., Sen, P.K., 1969. A Class of Rank Order Tests for a General Linear
Hypothesis. The Annals of Mathematical Statistics 40, 1325–1343.

Scheirer, C.J., Ray, W.S., Hare, N., 1976. The Analysis of Ranked Data
Derived from Completely Randomized Factorial Designs. Biometrics 32,
429–434. doi:10.2307/2529511
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[stats]{aov}}
}

