% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-calibration.R
\name{relerr}
\alias{relerr}
\title{Functions for weighted calibrations}
\usage{
relerr(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link[envalysis]{calibration}}.}

\item{\dots}{currently not used.}
}
\description{
ff
}
\examples{
data(din32645)
din <- calibration(Area ~ Conc, data = din32645)

relerr(din)

}
\seealso{
Other calibration: 
\code{\link{calibration}()}
}
\concept{calibration}
